+++ pfio/ClientManager.F90
@@ -1,6 +1,4 @@
-#include "MAPL_ErrLog.h"
-#include "unused_dummy.H"
-
+#include "MAPL.h"
 module pFIO_ClientManagerMod
 
    use mapl_ErrorHandling_mod
@@ -12,51 +10,23 @@ module pFIO_ClientManagerMod
    implicit none
    private
 
-   public :: init_IO_ClientManager
-   public :: get_client_thread
-   public :: client_map
+   public :: add_client
+   public :: get_client
 
-   interface init_IO_ClientManager
-      module procedure init_ClientManager
-   end interface
-
-   type(StringClientThreadMap), target, protected :: client_map
+   type(StringClientThreadMap), target, private :: client_map
 
 contains
 
-   subroutine init_ClientManager(client_comm, unusable, fast_oclient, rc)
-      integer, intent(in) :: client_comm
-      class (KeywordEnforcer), optional, intent(out) :: unusable
-      logical, optional, intent(in) :: fast_oclient
+   subroutine add_client(name, client, rc)
+      character(*), intent(in) :: name
+      class(ClientThread), intent(in) :: client
       integer, optional, intent(out) :: rc
-      integer :: status
 
-      logical :: fast_
-      type(ClientThread) :: i_client
-      class(ClientThread), allocatable :: o_client
-
-      fast_ = .false.
-      if (present(fast_oclient)) fast_ = fast_oclient
-
-      client_map = StringClientThreadMap()
-
-      i_client = ClientThread(client_comm=client_comm, rc=status)
-      _VERIFY(status)
-      call client_map%insert('i_client', i_client)
-
-      if (fast_) then
-         allocate(o_client, source=FastClientThread(client_comm=client_comm, rc=status))
-      else
-         allocate(o_client, source=ClientThread(client_comm=client_comm, rc=status))
-      end if
-      _VERIFY(status)
-      call client_map%insert('o_client', o_client)
-
+      call client_map%insert(name, client)
       _RETURN(_SUCCESS)
-      _UNUSED_DUMMY(unusable)
-   end subroutine init_ClientManager
+   end subroutine add_client
 
-   function get_client_thread(name, rc) result(client)
+   function get_client(name, rc) result(client)
       character(len=*), intent(in) :: name
       integer, optional, intent(out) :: rc
       class(ClientThread), pointer :: client
@@ -64,6 +34,6 @@ contains
       client => client_map%at(name)
       _ASSERT(associated(client), "Client '"//name//"' not found in client manager")
       _RETURN(_SUCCESS)
-   end function get_client_thread
+   end function get_client
 
 end module pFIO_ClientManagerMod