+++ pfio/ServerNames.F90
@@ -1,0 +1,7 @@
+module mapl_DefaultServerNames_mod
+   implicit none
+
+   character(*), parameter :: MAPL_DEFAULT_INPUT_SERVER  = 'mapl_default_input_server'
+   character(*), parameter :: MAPL_DEFAULT_OUTPUT_SERVER = 'mapl_default_output_server'
+
+end module mapl_DefaultServerNames_mod